# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AI Travel Planner API
  description: API for AI-powered travel planning application
  version: 1.0.0
  contact:
    name: API Support
    email: support@ai-travel-planner.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.ai-travel-planner.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /ml/nlp/parse:
    post:
      summary: Parse natural language travel request
      tags: [ML]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: "I want to visit Paris for 5 days with a budget of 2000 euros"
      responses:
        '200':
          description: Parsed travel request
          content:
            application/json:
              schema:
                type: object
                properties:
                  destination:
                    type: string
                  days:
                    type: number
                  budget:
                    type: number
                  interests:
                    type: array
                    items:
                      type: string
                  confidence:
                    type: object
  /export/pdf:
    post:
      summary: Export itinerary as PDF
      tags: [Export]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - itinerary_id
              properties:
                itinerary_id:
                  type: string
                include_map:
                  type: boolean
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary